Add previous line or `hunk.old_start` when a diff is line addition
Since added lines did not exist in previous file, the line number of
those cannot be used to determine which tests to run (since test-to-code
mapping is a mapping from line number in *previous file* to a test
file).

So, this commit changes to add lines *before* the added lines
(`prev_line`). If the addition starts from the first line of a file,
it adds `hunk.old_start` (which is most certainly the first line if
patches are generated the normal way).